+++ CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Documentation on the release process ([#241](https://github.com/stac-utils/stactools/pull/241))
+
 ## [0.2.6] - 2022-02-15
 
 ### Added
 
 - CI checks for minimum and pre-release versions of dependencies ([#228](https://github.com/stac-utils/stactools/pull/228))
 
+### Fixed
+
+- Test items are now valid STAC v1.0.0 ([#243](https://github.com/stac-utils/stactools/pull/243))
+- Move asset tests now have all the assets they need ([#243](https://github.com/stac-utils/stactools/pull/243))
 ### Changed
 
 - Use [pytest](https://docs.pytest.org/) for unit testing instead of `unittest` ([#220](https://github.com/stac-utils/stactools/pull/220))